fix: only create disk cache directory if disk cache and cache is enabled (#444)
* fix: only create disk cache directory if disk cache and cache is enabled
* fix: skip cache file cleanup task if directory doesn't exists

+++ orebfuscator-plugin/src/main/java/net/imprex/orebfuscator/cache/CacheFileCleanupTask.java
@@ -25,6 +25,11 @@ public class CacheFileCleanupTask implements Runnable {
 
 	@Override
 	public void run() {
+		if (Files.notExists(this.cacheConfig.baseDirectory())) {
+			OFCLogger.debug("Skipping CacheFileCleanupTask as the cache directory doesn't exist.");
+			return;
+		}
+
 		long deleteAfterMillis = this.cacheConfig.deleteRegionFilesAfterAccess();
 		AbstractRegionFileCache<?> regionFileCache = OrebfuscatorNms.getRegionFileCache();
